Fixed last tests, converted one test in phone number admin test to javascript test.
